regressiontest/requirements.txt
pytest==8.3.4
httpx[http2]==0.28.1
pytest-asyncio==0.24.0
playwright==1.50.0
python-dotenv==1.0.1
pytest-json-report==1.5.0
jinja2==3.1.5
anyio==4.7.0
